remove victory-continuous-transition component and let victory-transition do it all
